ml/object-detection/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
pillow>=10.2.0
numpy>=1.24.0
opencv-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
python-dotenv>=1.0.0
